rm old controller based watcher impl
