+++ go.work
@@ -3,5 +3,4 @@ go 1.19
 use (
 	./kcp
 	./skr
-	./skr/webhook
 )
