+++ skr/Dockerfile
@@ -1,29 +1,21 @@
-# Build the manager binary
-FROM golang:1.18 as builder
+FROM golang:1.19-alpine as builder
 
-WORKDIR /workspace
-# Copy the Go Modules manifests
+WORKDIR /app
+
 COPY go.mod go.mod
 COPY go.sum go.sum
-# cache deps before building and copying source so that we don't need to re-download as much
-# and so that source changes don't invalidate our downloaded layer
+
 RUN go mod download
 
-# Copy the go source
 COPY main.go main.go
-# Only needed if CRDs will be introduced
-# COPY api/ api/
-COPY controllers/ controllers/
-COPY pkg/ pkg/
+COPY internal/ internal/
 
-# Build
-RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -a -o manager main.go
+RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -a -o webhook main.go
 
-# Use distroless as minimal base image to package the manager binary
-# Refer to https://github.com/GoogleContainerTools/distroless for more details
 FROM gcr.io/distroless/static:nonroot
 WORKDIR /
-COPY --from=builder /workspace/manager .
+
+COPY --from=builder /app/webhook .
 USER 65532:65532
 
-ENTRYPOINT ["/manager"]
+ENTRYPOINT ["/webhook"]
