--- skr/chart/Chart.yaml
@@ -1,5 +1,0 @@
-apiVersion: v2
-description: A Helm chart for Kyma Watcher
-name: kyma-watcher
-type: application
-version: 0.1.0
