--- skr/chart/values.yaml
@@ -1,2 +1,0 @@
-## currently not used
-## -- might be used in the future
