--- skr/config/manager/kustomization.yaml
@@ -1,19 +1,0 @@
-resources:
-- manager.yaml
-
-generatorOptions:
-  disableNameSuffixHash: true
-
-configMapGenerator:
-- files:
-  - controller_manager_config.yaml
-  name: manager-config
-apiVersion: kustomize.config.k8s.io/v1beta1
-kind: Kustomization
-images:
-- name: controller
-  newName: /kyma-watcher
-- name: controller-manager
-  newName: docker.io/jaythedevil666/kyma-watcher
-- name: kyma-watcher
-  newName: k3d-registry.localhost:4000/kyma-watcher
