--- skr/config/prometheus/kustomization.yaml
@@ -1,2 +1,0 @@
-resources:
-- monitor.yaml
