+++ skr/internal/assets/configmap.yaml
