+++ skr/main.go
@@ -1,116 +1,90 @@
-/*
-Copyright 2022.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
 package main
 
 import (
 	"flag"
-	v1 "k8s.io/api/core/v1"
+	"net/http"
 	"os"
+	"strconv"
 
-	// Import all Kubernetes client auth plugins (e.g. Azure, GCP, OIDC, etc.)
-	// to ensure that exec-entrypoint and run can make use of them.
-	_ "k8s.io/client-go/plugin/pkg/client/auth"
+	"github.com/kyma-project/runtime-watcher/skr/internal"
 
-	"k8s.io/apimachinery/pkg/runtime"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
-	clientgoscheme "k8s.io/client-go/kubernetes/scheme"
 	ctrl "sigs.k8s.io/controller-runtime"
-	"sigs.k8s.io/controller-runtime/pkg/healthz"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
-
-	"github.com/kyma-project/kyma-watcher/skr/controllers"
-	"github.com/kyma-project/kyma-watcher/skr/pkg/config"
-	//+kubebuilder:scaffold:imports
 )
 
-var (
-	scheme   = runtime.NewScheme()        //nolint:gochecknoglobals
-	setupLog = ctrl.Log.WithName("setup") //nolint:gochecknoglobals
-)
-
-const (
-	port = 9443
-)
-
-func init() { //nolint:gochecknoinits
-	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
-	utilruntime.Must(v1.AddToScheme(scheme))
-
-	//+kubebuilder:scaffold:scheme
+type ServerParameters struct {
+	port       int    // webhook server port
+	certFile   string // path to TLS certificate for https
+	keyFile    string // path to TLS key matching for certificate
+	tlsEnabled bool   // indicates if TLS is enabled
 }
 
-func main() {
-	var metricsAddr string
-	var enableLeaderElection bool
-	var probeAddr string
-	var kcpIP, kcpPort string
+var parameters ServerParameters //nolint:gochecknoglobals
 
-	flag.StringVar(&metricsAddr, "metrics-bind-address", ":8083", "The address the metric endpoint binds to.")
-	flag.StringVar(&probeAddr, "health-probe-bind-address", ":8084", "The address the probe endpoint binds to.")
-	flag.BoolVar(&enableLeaderElection, "leader-elect", false,
-		"Enable leader election for controller manager. "+
-			"Enabling this will ensure there is only one active controller manager.")
-	flag.StringVar(&kcpIP, "kcp-ip", config.KcpIP, "IP-Address of KCP")
-	flag.StringVar(&kcpPort, "kcp-port", config.KcpPort, "Exposed event port of KCP")
+func main() {
+	logger := ctrl.Log.WithName("skr-webhook")
+
 	opts := zap.Options{
 		Development: true,
 	}
+
 	opts.BindFlags(flag.CommandLine)
 	flag.Parse()
 
 	ctrl.SetLogger(zap.New(zap.UseFlagOptions(&opts)))
 
-	mgr, err := ctrl.NewManager(ctrl.GetConfigOrDie(), ctrl.Options{
-		Scheme:                 scheme,
-		MetricsBindAddress:     metricsAddr,
-		Port:                   port,
-		HealthProbeBindAddress: probeAddr,
-		LeaderElection:         enableLeaderElection,
-		LeaderElectionID:       "893110f8.kyma-project.io",
-	})
+	var err error
+
+	// port
+	portEnv := os.Getenv("WEBHOOK_PORT")
+	defaultPort := 8443
+	if portEnv != "" {
+		defaultPort, err = strconv.Atoi(portEnv)
+		if err != nil {
+			logger.Error(err, "Error parsing Webhook server port")
+		}
+	}
+	parameters.port = defaultPort
+
+	// tls
+	tlsEnabledEnv := os.Getenv("TLS_ENABLED")
+	tlsEnabled := false
+	if tlsEnabledEnv != "" {
+		tlsEnabled, err = strconv.ParseBool(tlsEnabledEnv)
+		if err != nil {
+			logger.Error(err, "Error parsing tls flag")
+		}
+	}
+	parameters.tlsEnabled = tlsEnabled
+	parameters.certFile = os.Getenv("TLS_CERT")
+	parameters.keyFile = os.Getenv("TLS_KEY")
+
+	restConfig := ctrl.GetConfigOrDie()
+
+	restClient, err := client.New(restConfig, client.Options{})
 	if err != nil {
-		setupLog.Error(err, "unable to start manager")
-		os.Exit(1)
+		logger.Error(err, "rest client could not be determined for skr-webhook")
+		return
 	}
 
-	if err = (&controllers.KymaWatcherReconciler{
-		Client:  mgr.GetClient(),
-		Scheme:  mgr.GetScheme(),
-		Logger:  mgr.GetLogger(),
-		KcpIP:   kcpIP,
-		KcpPort: kcpPort,
-	}).SetupWithManager(mgr); err != nil {
-		setupLog.Error(err, "unable to create controller", "controller", "KymaWatcherReconciler")
-		os.Exit(1)
+	// handler
+	handler := &internal.Handler{
+		Client: restClient,
+		Logger: logger,
 	}
+	http.HandleFunc("/validate/", handler.Handle)
 
-	//+kubebuilder:scaffold:builder
-	if err := mgr.AddHealthzCheck("healthz", healthz.Ping); err != nil {
-		setupLog.Error(err, "unable to set up health check")
-		os.Exit(1)
+	// server
+	logger.Info("starting web server", "Port:", parameters.port)
+	if parameters.tlsEnabled {
+		err = http.ListenAndServeTLS(":"+strconv.Itoa(parameters.port), parameters.certFile,
+			parameters.keyFile, nil)
+	} else {
+		err = http.ListenAndServe(":"+strconv.Itoa(parameters.port), nil)
 	}
-	if err := mgr.AddReadyzCheck("readyz", healthz.Ping); err != nil {
-		setupLog.Error(err, "unable to set up ready check")
-		os.Exit(1)
-	}
-
-	setupLog.Info("starting manager")
-	if err := mgr.Start(ctrl.SetupSignalHandler()); err != nil {
-		setupLog.Error(err, "problem running manager")
-		os.Exit(1)
+	if err != nil {
+		logger.Error(err, "error starting skr-webhook server")
+		return
 	}
 }
