--- skr/pkg/config/values.go
@@ -1,18 +1,0 @@
-package config
-
-// Default values.
-const ComponentLabel = "operator.kyma-project.io/controller-name"
-const KymaCrLabel = "operator.kyma-project.io/kyma-name"
-
-const (
-	WatcherSecretName      = "sample-secret"
-	WatcherSecretNamespace = "operator-system"
-)
-
-const (
-	KcpIP           = "http://localhost"
-	KcpPort         = "8082"
-	ContractVersion = "v1"
-)
-
-const EventEndpoint = "event"
