--- skr/webhook/Dockerfile
@@ -1,21 +1,0 @@
-FROM golang:1.19-alpine as builder
-
-WORKDIR /app
-
-COPY go.mod go.mod
-COPY go.sum go.sum
-
-RUN go mod download
-
-COPY main.go main.go
-COPY internal/ internal/
-
-RUN CGO_ENABLED=0 GOOS=linux GOARCH=amd64 go build -a -o webhook main.go
-
-FROM gcr.io/distroless/static:nonroot
-WORKDIR /
-
-COPY --from=builder /app/webhook .
-USER 65532:65532
-
-ENTRYPOINT ["/webhook"]
